Edited GameListActivity and added toasts/exception handling

+++ app/src/main/java/edu/orangecoastcollege/cs273/mpaulding/gamersdelight/GameListActivity.java
@@ -7,8 +7,12 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.RatingBar;
+import android.widget.Toast;
 
 import java.util.List;
+
+import static edu.orangecoastcollege.cs273.mpaulding.gamersdelight.R.id.descriptionEditText;
+import static edu.orangecoastcollege.cs273.mpaulding.gamersdelight.R.id.nameEditText;
 
 public class GameListActivity extends AppCompatActivity {
 
@@ -17,9 +21,10 @@ public class GameListActivity extends AppCompatActivity {
     private GameListAdapter gamesListAdapter;
     private ListView gamesListView;
 
-    private EditText nameEditText;
-    private EditText descriptionEditText;
+    private EditText gameNameEditText;
+    private EditText gameDescriptionEditText;
     private RatingBar gameRatingBar;
+    //private ImageView gameImageView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -35,9 +40,9 @@ public class GameListActivity extends AppCompatActivity {
         db.addGame(new Game("Doom FLH", "First person shooter", 2.5f, "doomflh.png"));
         db.addGame(new Game("Battlefield 1", "Single player campaign", 5.0f, "battlefield1.png"));
 
-        nameEditText = (EditText) findViewById(R.id.nameEditText);
-        descriptionEditText = (EditText) findViewById(R.id.descriptionEditText);
-        gameRatingBar = (RatingBar) findViewById(R.id.gameListRatingBar);
+        gameNameEditText = (EditText) findViewById(nameEditText);
+        gameDescriptionEditText = (EditText) findViewById(descriptionEditText);
+        gameRatingBar = (RatingBar) findViewById(R.id.gameRatingBar);
 
         // TODO:  Populate all games from the database into the list
         gamesList = db.getAllGames();
@@ -53,10 +58,20 @@ public class GameListActivity extends AppCompatActivity {
         // TODO: Use an Intent to start the GameDetailsActivity with the data it needs to correctly inflate its views.
         Intent gamesListIntent = new Intent(this, GameDetailsActivity.class);
 
-        gamesListIntent.putExtra("", "Name");
-        gamesListIntent.putExtra("", "Description");
-        gamesListIntent.putExtra("", "Rating");
-        gamesListIntent.putExtra("", "Image");
+        String gameName = gameNameEditText.getText().toString();
+        String gameDescription = gameDescriptionEditText.getText().toString();
+        String gameRating = String.valueOf(gameRatingBar.getRating());
+        String gameImageName = "avatar.png";
+
+        //gamesListIntent.putExtra("", "Name");
+        //gamesListIntent.putExtra("", "Description");
+        //gamesListIntent.putExtra(String.valueOf(0.0f), "Rating");
+        //gamesListIntent.putExtra("avatar.png", "Image");
+
+        gamesListIntent.putExtra(gameName, "Name");
+        gamesListIntent.putExtra(gameDescription, "Description");
+        gamesListIntent.putExtra(gameRating, "Rating");
+        gamesListIntent.putExtra(gameImageName, "Image");
 
         startActivity(gamesListIntent);
     }
@@ -64,21 +79,45 @@ public class GameListActivity extends AppCompatActivity {
     public void addGame(View view)
     {
         // TODO:  Add a game to the database, list, list adapter
-        Game newGame = new Game("", "", 0.0f, "avatar.png");
-        gamesListAdapter.add(newGame);
-        // TODO:  Make sure the list adapter is updated
-        gamesListAdapter.notifyDataSetChanged();
-        // TODO:  Clear all entries the user made (edit text and rating bar)
-        nameEditText.setText("");
-        descriptionEditText.setText("");
-        gameRatingBar.setRating(0.0f);
+        //gameNameEditText = (EditText) findViewById(nameEditText);
+        //gameDescriptionEditText = (EditText) findViewById(descriptionEditText);
+        //gameRatingBar = (RatingBar) findViewById(R.id.gameRatingBar);
+
+        String gameName = gameNameEditText.getText().toString();
+        if (gameName.isEmpty())
+            Toast.makeText(this, "Game name cannot be empty.", Toast.LENGTH_SHORT).show();
+        else {
+            String gameDescription = gameDescriptionEditText.getText().toString();
+            if (gameDescription.isEmpty())
+                Toast.makeText(this, "Game description cannot be empty.", Toast.LENGTH_SHORT).show();
+            else {
+                float gameRating = gameRatingBar.getRating();
+                String gameImageName = "avatar.png";
+
+                //Game newGame = new Game("", "", 0.0f, "avatar.png");
+                Game newGame = new Game(gameName, gameDescription, gameRating, gameImageName);
+                gamesListAdapter.add(newGame);
+
+                // TODO:  Make sure the list adapter is updated
+                gamesListAdapter.notifyDataSetChanged();
+
+                // TODO:  Clear all entries the user made (edit text and rating bar)
+                gameNameEditText.setText("");
+                gameDescriptionEditText.setText("");
+                gameRatingBar.setRating(0.0f);
+            }
+        }
     }
 
     public void clearAllGames(View view)
     {
         // TODO:  Delete all games from the database and lists
-        gamesList.clear();
-        db.deleteAllGames();
-        gamesListAdapter.notifyDataSetChanged();
+        if (!gamesList.isEmpty()) {
+            gamesList.clear();
+            db.deleteAllGames();
+            gamesListAdapter.notifyDataSetChanged();
+        }
+        else
+            Toast.makeText(this, "Game list is already empty.", Toast.LENGTH_SHORT).show();
     }
 }
